Added quick fixes for focus and footer
